Some c2ctool upgrade improvements
Don't commit the old folder.
Do the diff in a file.
Ask for a buildout file to use.
Automatically test the check collector.
Add missing version in the doc.
Add a mobile application title in the template params
Verify [Version] of c2cgeoportal and develop flag
Don't do a push after an upgrade